modifi

+++ app.py
@@ -5,7 +5,6 @@ import secrets
 import os
 from pathlib import Path
 from config import Config
-
 from database.models import create_tables
 from routes.login import login_bp
 from routes.register import register_bp
@@ -14,13 +13,7 @@ from routes.check_username import check_username_bp
 from routes.admin import admin_bp
 from routes.simulation import simulation_bp
 from database.db import get_db, close_db
-
-
-
-
-
-
-
+from routes.compound import compound_bp
 
 
 app = Flask(__name__)
@@ -31,6 +24,7 @@ app.register_blueprint(check_email_bp)
 app.register_blueprint(check_username_bp)
 app.register_blueprint(admin_bp)
 app.register_blueprint(simulation_bp)
+app.register_blueprint(compound_bp)
 
 
 
@@ -58,42 +52,7 @@ def go_home():
     # هنا نستخدم اسم الدالة home، وليس اسم الملف
     return redirect(url_for('home'))
 
-@app.route('/forgot', methods=["GET", "POST"])
-def forgot():
-    if request.method == 'POST':
-        user_email = request.form['email']
-        token = secrets.token_urlsafe(16)
-        reset_link = url_for('reset_password', token=token, _external=True)
 
-        msg = Message("استرجاع كلمة السر",
-                      sender="kindkiki9@gmail.com",
-                      recipients=[user_email])
-        msg.body = f"اضغطي على الرابط التالي لإعادة تعيين كلمة السر:\n{reset_link}"
-
-        try:
-            mail.send(msg)
-            flash("تم إرسال رابط إعادة تعيين كلمة السر إلى بريدك الإلكتروني")
-        except Exception as e:
-            print(f" خطأ في إرسال البريد: {e}")
-            flash("حدث خطأ أثناء إرسال البريد، حاول لاحقاً")
-
-        return redirect(url_for('forgot'))
-
-    return render_template('login/forgot.html')
-
-@app.route('/reset-password/<token>', methods=['GET', 'POST'])
-def reset_password(token):
-    if request.method == 'POST':
-        new_password = request.form['password']
-        flash("تم تحديث كلمة السر بنجاح!")
-        return redirect(url_for('login.login'))
-
-    return """
-    <form method="POST">
-        <input type="password" name="password" placeholder="New Password" required>
-        <button type="submit">Reset Password</button>
-    </form>
-    """
 
 @app.route('/profile')
 def profile():
@@ -105,6 +64,7 @@ def profile():
     cursor = db.cursor()
     cursor.execute("SELECT * FROM user WHERE id = ?", (user_id,))
     user = cursor.fetchone()  # يفترض أنك تستخدم row_factory = sqlite3.Row
+    db.close()
     if not user:
         flash("المستخدم غير موجود")
         return redirect(url_for("login.login"))
@@ -136,10 +96,10 @@ def update_profile():
         # حذف الصورة القديمة إذا لم تكن default
         cursor.execute("SELECT avatar FROM user WHERE id = ?", (user_id,))
         old_avatar = cursor.fetchone()["avatar"]
-        if old_avatar != 'default.png':
-            old_path = Path(app.config['UPLOAD_FOLDER']) / old_avatar
-            if old_path.exists():
-                old_path.unlink()
+        if old_avatar and old_avatar != 'default.png':
+          old_path = Path(app.config['UPLOAD_FOLDER']) / old_avatar
+          if old_path.exists():
+           old_path.unlink()
 
     # تحديث قاعدة البيانات
     if username and avatar_filename:
@@ -149,10 +109,48 @@ def update_profile():
     elif avatar_filename:
         cursor.execute("UPDATE user SET avatar = ? WHERE id = ?", (avatar_filename, user_id))
 
+    if username:
+        cursor.execute("SELECT id FROM user WHERE username = ?", (username,))
+        existing_user = cursor.fetchone()
+
+        if existing_user and existing_user["id"] != user_id:
+            flash("Username already exists!", "error")
+            return redirect(url_for('profile'))
+
     db.commit()
     flash("تم تحديث البروفايل بنجاح!", "success")
     return redirect(url_for('profile'))
 
+
+
+@app.route('/delete-avatar', methods=['POST'])
+def delete_avatar():
+    user_id = session.get("user_id")
+    if not user_id:
+        return redirect(url_for("login.login"))
+
+    db = get_db()
+    cursor = db.cursor()
+
+    # جلب الصورة الحالية
+    cursor.execute("SELECT avatar FROM user WHERE id = ?", (user_id,))
+    user = cursor.fetchone()
+
+    if user and user["avatar"]:
+        avatar = user["avatar"]
+
+        # حذف الصورة من السيرفر
+        avatar_path = Path(app.config['UPLOAD_FOLDER']) / avatar
+        if avatar_path.exists():
+            avatar_path.unlink()
+
+        # حذفها من قاعدة البيانات
+        cursor.execute("UPDATE user SET avatar = NULL WHERE id = ?", (user_id,))
+        db.commit()
+
+        flash("Avatar deleted successfully", "success")
+
+    return redirect(url_for('profile'))
 
 
 @app.route('/search')
@@ -171,12 +169,30 @@ def settings():
 @app.route('/logout')
 def logout():
     session.clear()
-    return redirect(url_for('home'))
+    return redirect(url_for('login.login'))
 
 
 
+def get_compound_data(name):
+    url = f"https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/name/{name}/JSON"
+    response = requests.get(url)
 
+    if response.status_code != 200:
+        return None
 
+    data = response.json()
+
+    try:
+        info = data['PC_Compounds'][0]
+
+        return {
+            "name": name,
+            "cid": info.get("id", {}).get("id", {}).get("cid"),
+        }
+    except:
+        return None
+    
+   
 if __name__ == "__main__":
     print("بدء تشغيل التطبيق...")
